Fix for bug 59621: Line width increases to 1000mm
svn path=/trunk/koffice/; revision=230860

+++ kivio/kiviopart/KIvioViewIface.cc
@@ -195,7 +195,7 @@ void KIvioViewIface::unselectAllStencils()
     view->unselectAllStencils();
 }
 
-int KIvioViewIface::lineWidth() const
+double KIvioViewIface::lineWidth() const
 {
     return view->lineWidth();
 }
